Move rucio-dataset-monitoring to parent directory
