+++ rucio-dataset-monitoring/controllers/detailed_datasets.go
@@ -4,9 +4,9 @@ package controllers
 
 import (
 	"context"
-	"github.com/dmwm/CMSMonitoring/src/go/rucio-dataset-monitoring/models"
-	mymongo "github.com/dmwm/CMSMonitoring/src/go/rucio-dataset-monitoring/mongo"
-	"github.com/dmwm/CMSMonitoring/src/go/rucio-dataset-monitoring/utils"
+	"github.com/dmwm/CMSMonitoring/rucio-dataset-monitoring/models"
+	mymongo "github.com/dmwm/CMSMonitoring/rucio-dataset-monitoring/mongo"
+	"github.com/dmwm/CMSMonitoring/rucio-dataset-monitoring/utils"
 	"github.com/gin-gonic/gin"
 	"net/http"
 )
