+++ rucio-dataset-monitoring/static/templates/detailed_datasets.tmpl
